+++ src/components/Cartpage/DeleteCart.js
@@ -1,6 +1,13 @@
 import { HiX } from "react-icons/hi";
 import { useState } from 'react';
 import axios from "axios";
+import styled from 'styled-components';
+
+const Delete = styled.p`
+  display: flex;
+  align-self: flex-end;
+  margin: 0 268px 0 0;
+`
 
 function DeleteCart() {
   const AllDelete = () => {
@@ -15,17 +22,17 @@ function DeleteCart() {
     // 성공 핸들링
     if(response.status == 204) {
       alert('장바구니의 모든 상품이 삭제되었습니다.')
+      window.location.reload();
     }
   })
   .catch(function (error) {
     // 에러 핸들링
     console.log(error);
   })
-
 }
 
     return(
-        <p onClick={AllDelete}>전체삭제</p>
+        <Delete onClick={AllDelete}>전체삭제</Delete>
     )
 }
