+++ src/index.css
@@ -20,7 +20,9 @@ html {
 
 body {
   width: 100vw;
+  height: 100vh;
   background-color: #fff;
+  overflow-x: hidden;
 }
 
 a { text-decoration: none;}
